+++ api/v1/status/status.go
@@ -3,6 +3,8 @@ package status
 import (
 	"reflect"
 
+	"github.com/10gen/ops-manager-kubernetes/api/v1/status/pvc"
+
 	"github.com/10gen/ops-manager-kubernetes/pkg/util/stringutil"
 	"github.com/10gen/ops-manager-kubernetes/pkg/util/timeutil"
 )
@@ -25,6 +27,7 @@ type Reader interface {
 	// GetStatus returns the status of the object. The list of Options
 	// provided indicates which subresource will be returned. AppDB, OM or Backup
 	GetStatus(options ...Option) interface{}
+	GetCommonStatus(options ...Option) *Common
 }
 
 // Common is the struct shared by all statuses in existing Custom Resources.
@@ -35,6 +38,41 @@ type Common struct {
 	LastTransition     string             `json:"lastTransition,omitempty"`
 	ResourcesNotReady  []ResourceNotReady `json:"resourcesNotReady,omitempty"`
 	ObservedGeneration int64              `json:"observedGeneration,omitempty"`
+	PVCs               PVCS               `json:"pvc,omitempty"`
+}
+
+type PVCS []PVC
+
+func (p *PVCS) Merge(pvc2 PVC) PVCS {
+	if p == nil {
+		return nil
+	}
+
+	found := false
+	for i := range *p {
+		if (*p)[i].StatefulsetName == pvc2.StatefulsetName {
+			found = true
+			(*p)[i].Phase = pvc2.Phase
+		}
+	}
+
+	if !found {
+		*p = append(*p, pvc2)
+	}
+
+	return *p
+}
+
+type PVC struct {
+	Phase           pvc.Phase `json:"phase"`
+	StatefulsetName string    `json:"statefulsetName"`
+}
+
+func (p *PVC) GetPhase() pvc.Phase {
+	if p == nil || p.StatefulsetName == "" || p.Phase == "" {
+		return pvc.PhaseNoAction
+	}
+	return p.Phase
 }
 
 // UpdateCommonFields is the update function to update common fields used in statuses of all managed CRs
@@ -47,6 +85,14 @@ func (s *Common) UpdateCommonFields(phase Phase, generation int64, statusOptions
 	}
 	if option, exists := GetOption(statusOptions, ResourcesNotReadyOption{}); exists {
 		s.ResourcesNotReady = option.(ResourcesNotReadyOption).ResourcesNotReady
+	}
+	if option, exists := GetOption(statusOptions, PVCStatusOption{}); exists {
+		p := option.(PVCStatusOption).PVC
+		if p == nil {
+			s.PVCs = nil
+		} else {
+			s.PVCs.Merge(*p)
+		}
 	}
 	// We update the time only if the status really changed. Otherwise, we'd like to preserve the old one.
 	if !reflect.DeepEqual(previousStatus, s) {